2009-02-13  Jeffrey Stedfast  <fejj@novell.com>
	* mediaelement.h (class MediaElement): NaturalVideoWidth/Height
	should be ints, not doubles.


svn path=/trunk/moon/; revision=126887

+++ src/mediaelement.cpp
@@ -575,8 +575,8 @@ MediaElement::SetMedia (Media *media)
 	media->SetBufferingEnabled (true);
 	
 	SetNaturalDuration (broadcast ? 0 : TimeSpan_FromPts (mplayer->GetDuration ()));
-	SetNaturalVideoHeight ((double) mplayer->GetVideoHeight ());
-	SetNaturalVideoWidth ((double) mplayer->GetVideoWidth ());
+	SetNaturalVideoHeight (mplayer->GetVideoHeight ());
+	SetNaturalVideoWidth (mplayer->GetVideoWidth ());
 	SetAudioStreamCount (mplayer->GetAudioStreamCount ());
 	
 	if (mplayer->HasAudio ()) {
